+++ test/unit/chat_hide_test.dart
@@ -101,6 +101,7 @@ void main() async {
     'members': {'nodes': [], 'totalCount': 0},
     'kind': 'GROUP',
     'isHidden': false,
+    'isArchived': false,
     'muted': null,
     'directLink': null,
     'createdAt': '2021-12-15T15:11:18.316846+00:00',
@@ -161,6 +162,9 @@ void main() async {
     graphQlProvider.recentChatsTopEvents(3),
   ).thenAnswer((_) => const Stream.empty());
   when(
+    graphQlProvider.recentChatsTopEvents(3, archived: true),
+  ).thenAnswer((_) => const Stream.empty());
+  when(
     graphQlProvider.incomingCallsTopEvents(3),
   ).thenAnswer((_) => const Stream.empty());
   when(
@@ -194,6 +198,7 @@ void main() async {
         last: null,
         before: null,
         noFavorite: anyNamed('noFavorite'),
+        archived: anyNamed('archived'),
         withOngoingCalls: anyNamed('withOngoingCalls'),
       ),
     ).thenAnswer((_) => Future.value(RecentChats$Query.fromJson(recentChats)));
@@ -307,6 +312,7 @@ void main() async {
         last: null,
         before: null,
         noFavorite: anyNamed('noFavorite'),
+        archived: anyNamed('archived'),
         withOngoingCalls: anyNamed('withOngoingCalls'),
       ),
     ).thenAnswer((_) => Future.value(RecentChats$Query.fromJson(recentChats)));