+++ test/widget/chat_reply_message_test.dart
@@ -111,6 +111,9 @@ void main() async {
       ),
     ),
   );
+  when(
+    graphQlProvider.recentChatsTopEvents(3, archived: true),
+  ).thenAnswer((_) => const Stream.empty());
 
   final StreamController<QueryResult> contactEvents = StreamController();
   when(
@@ -214,6 +217,7 @@ void main() async {
       last: null,
       before: null,
       noFavorite: anyNamed('noFavorite'),
+      archived: anyNamed('archived'),
       withOngoingCalls: anyNamed('withOngoingCalls'),
     ),
   ).thenAnswer((_) => Future.value(RecentChats$Query.fromJson(recentChats)));
@@ -557,6 +561,7 @@ void main() async {
     expect(find.richText('reply message', skipOffstage: false), findsOneWidget);
 
     await Future.wait([common.close(), scoped.close()]);
+
     await Get.deleteAll(force: true);
   });
 }
@@ -582,6 +587,7 @@ final chatData = {
   'members': {'nodes': [], 'totalCount': 0},
   'kind': 'GROUP',
   'isHidden': false,
+  'isArchived': false,
   'muted': null,
   'directLink': null,
   'createdAt': '2021-12-15T15:11:18.316846+00:00',